web_app/src/server/agent/python/examples/cdp-langchain-chatbot/pyproject.toml
[tool.poetry]
name = "chatbot-python"
version = "0.0.1"
description = "CDP AgentKit Example Chatbot"
authors = ["John Peterson <john.peterson@coinbase.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
cdp-langchain = "^0.0.13"
sounddevice = "^0.5.1"
langchain-community = "^0.3.16"
soundfile = "^0.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
